Move utility code to atomic_utils namespace.
