+++ searchlib/src/vespa/searchlib/attribute/load_utils.h
@@ -2,6 +2,7 @@
 
 #pragma once
 
+#include "atomic_utils.h"
 #include "attributevector.h"
 #include "readerbase.h"
 #include <vespa/vespalib/util/arrayref.h>
@@ -14,30 +15,6 @@ class EntryRef;
 }
 
 namespace search::attribute {
-
-namespace load_utils {
-
-/*
- * Helper class to map from atomic value to non-atomic value, e.g.
- * from AtomicEntryRef to EntryRef.
- */
-template <typename MaybeAtomicValue>
-class NonAtomicValue {
-public:
-    using type = MaybeAtomicValue;
-};
-
-template <>
-class NonAtomicValue<vespalib::datastore::AtomicEntryRef>
-{
-public:
-    using type = vespalib::datastore::EntryRef;
-};
-
-template <class MaybeAtomicValue>
-using NonAtomicValue_t = typename NonAtomicValue<MaybeAtomicValue>::type;
-
-}
 
 /**
  * Helper functions used to open / load attribute vector data files from disk.
@@ -82,7 +59,7 @@ void
 loadFromEnumeratedSingleValue(Vector &vector,
                               vespalib::GenerationHolder &genHolder,
                               ReaderBase &attrReader,
-                              vespalib::ConstArrayRef<load_utils::NonAtomicValue_t<typename Vector::ValueType>> enumValueToValueMap,
+                              vespalib::ConstArrayRef<atomic_utils::NonAtomicValue_t<typename Vector::ValueType>> enumValueToValueMap,
                               vespalib::ConstArrayRef<uint32_t> enum_value_remapping,
                               Saver saver) __attribute((noinline));
